@@ -24,28 +24,9 @@ import threading
 import time
 from abc import ABC
 from collections import defaultdict
+from collections.abc import Generator, Iterable, Iterator, Mapping, Sequence
 from sqlite3 import Connection
-from types import TracebackType
-from typing import (
-    Any,
-    AnyStr,
-    Callable,
-    DefaultDict,
-    Dict,
-    Generator,
-    Generic,
-    Iterable,
-    Iterator,
-    List,
-    Mapping,
-    Optional,
-    Sequence,
-    Tuple,
-    Type,
-    TypeVar,
-    Union,
-    cast,
-)
+from typing import TYPE_CHECKING, Any, AnyStr, Callable, Generic, TypeVar, cast
 
 from unidecode import unidecode
 
@@ -62,6 +43,9 @@ from .query import (
     Sort,
     TrueQuery,
 )
+
+if TYPE_CHECKING:
+    from types import TracebackType
 
 
 class DBAccessError(Exception):
@@ -120,7 +104,7 @@ class FormattedMapping(Mapping[str, str]):
     def get(  # type: ignore
         self,
         key: str,
-        default: Optional[str] = None,
+        default: str | None = None,
     ) -> str:
         """Similar to Mapping.get(key, default), but always formats to str."""
         if default is None:
@@ -158,14 +142,14 @@ class FormattedMapping(Mapping[str, str]):
 class LazyConvertDict:
     """Lazily convert types for attributes fetched from the database"""
 
-    def __init__(self, model_cls: "Model"):
+    def __init__(self, model_cls: Model):
         """Initialize the object empty"""
         # FIXME: Dict[str, SQLiteType]
-        self._data: Dict[str, Any] = {}
+        self._data: dict[str, Any] = {}
         self.model_cls = model_cls
-        self._converted: Dict[str, Any] = {}
+        self._converted: dict[str, Any] = {}
 
-    def init(self, data: Dict[str, Any]):
+    def init(self, data: dict[str, Any]):
         """Set the base data that should be lazily converted"""
         self._data = data
 
@@ -195,7 +179,7 @@ class LazyConvertDict:
         if key in self._data:
             del self._data[key]
 
-    def keys(self) -> List[str]:
+    def keys(self) -> list[str]:
         """Get a list of available field names for this object."""
         return list(self._converted.keys()) + list(self._data.keys())
 
@@ -213,14 +197,14 @@ class LazyConvertDict:
         for key, value in values.items():
             self[key] = value
 
-    def items(self) -> Iterable[Tuple[str, Any]]:
+    def items(self) -> Iterable[tuple[str, Any]]:
         """Iterate over (key, value) pairs that this object contains.
         Computed fields are not included.
         """
         for key in self:
             yield key, self[key]
 
-    def get(self, key: str, default: Optional[Any] = None):
+    def get(self, key: str, default: Any | None = None):
         """Get the value for a given key or `default` if it does not
         exist.
         """
@@ -286,7 +270,7 @@ class Model(ABC):
     """The flex field SQLite table name.
     """
 
-    _fields: Dict[str, types.Type] = {}
+    _fields: dict[str, types.Type] = {}
     """A mapping indicating available "fixed" fields on this type. The
     keys are field names and the values are `Type` objects.
     """
@@ -296,16 +280,16 @@ class Model(ABC):
     terms.
     """
 
-    _types: Dict[str, types.Type] = {}
+    _types: dict[str, types.Type] = {}
     """Optional Types for non-fixed (i.e., flexible and computed) fields.
     """
 
-    _sorts: Dict[str, Type[Sort]] = {}
+    _sorts: dict[str, type[Sort]] = {}
     """Optional named sort criteria. The keys are strings and the values
     are subclasses of `Sort`.
     """
 
-    _queries: Dict[str, Type[FieldQuery]] = {}
+    _queries: dict[str, type[FieldQuery]] = {}
     """Named queries that use a field-like `name:value` syntax but which
     do not relate to any specific field.
     """
@@ -348,7 +332,7 @@ class Model(ABC):
         return cls._relation._fields.keys() - cls.shared_db_fields
 
     @classmethod
-    def _getters(cls: Type["Model"]):
+    def _getters(cls: type[Model]):
         """Return a mapping from field names to getter functions."""
         # We could cache this if it becomes a performance problem to
         # gather the getter mapping every time.
@@ -363,7 +347,7 @@ class Model(ABC):
 
     # Basic operation.
 
-    def __init__(self, db: Optional[Database] = None, **values):
+    def __init__(self, db: Database | None = None, **values):
         """Create a new object with an optional Database association and
         initial field values.
         """
@@ -378,10 +362,10 @@ class Model(ABC):
 
     @classmethod
     def _awaken(
-        cls: Type[AnyModel],
-        db: Optional[Database] = None,
-        fixed_values: Dict[str, Any] = {},
-        flex_values: Dict[str, Any] = {},
+        cls: type[AnyModel],
+        db: Database | None = None,
+        fixed_values: dict[str, Any] = {},
+        flex_values: dict[str, Any] = {},
     ) -> AnyModel:
         """Create an object with values drawn from the database.
 
@@ -421,7 +405,7 @@ class Model(ABC):
 
         return self._db
 
-    def copy(self) -> "Model":
+    def copy(self) -> Model:
         """Create a copy of the model object.
 
         The field values and other state is duplicated, but the new copy
@@ -537,7 +521,7 @@ class Model(ABC):
         for key, value in values.items():
             self[key] = value
 
-    def items(self) -> Iterator[Tuple[str, Any]]:
+    def items(self) -> Iterator[tuple[str, Any]]:
         """Iterate over (key, value) pairs that this object contains.
         Computed fields are not included.
         """
@@ -579,7 +563,7 @@ class Model(ABC):
 
     # Database interaction (CRUD methods).
 
-    def store(self, fields: Optional[Iterable[str]] = None):
+    def store(self, fields: Iterable[str] | None = None):
         """Save the object's metadata into the library database.
         :param fields: the fields to be stored. If not specified, all fields
         will be.
@@ -653,7 +637,7 @@ class Model(ABC):
                 f"DELETE FROM {self._flex_table} WHERE entity_id=?", (self.id,)
             )
 
-    def add(self, db: Optional["Database"] = None):
+    def add(self, db: Database | None = None):
         """Add the object to the library database. This object must be
         associated with a database; you can provide one via the `db`
         parameter or use the currently associated database.
@@ -692,7 +676,7 @@ class Model(ABC):
 
     def evaluate_template(
         self,
-        template: Union[str, functemplate.Template],
+        template: str | functemplate.Template,
         for_path: bool = False,
     ) -> str:
         """Evaluate a template (a string or a `Template` object) using
@@ -730,16 +714,16 @@ class Model(ABC):
         cls,
         field,
         pattern,
-        query_cls: Type[FieldQuery] = MatchQuery,
+        query_cls: type[FieldQuery] = MatchQuery,
     ) -> FieldQuery:
         """Get a `FieldQuery` for this model."""
         return query_cls(field, pattern, field in cls._fields)
 
     @classmethod
     def all_fields_query(
-        cls: Type["Model"],
+        cls: type[Model],
         pats: Mapping,
-        query_cls: Type[FieldQuery] = MatchQuery,
+        query_cls: type[FieldQuery] = MatchQuery,
     ):
         """Get a query that matches many fields with different patterns.
 
@@ -764,11 +748,11 @@ class Results(Generic[AnyModel]):
 
     def __init__(
         self,
-        model_class: Type[AnyModel],
-        rows: List[Mapping],
-        db: "Database",
+        model_class: type[AnyModel],
+        rows: list[Mapping],
+        db: Database,
         flex_rows,
-        query: Optional[Query] = None,
+        query: Query | None = None,
         sort=None,
     ):
         """Create a result set that will construct objects of type
@@ -800,7 +784,7 @@ class Results(Generic[AnyModel]):
 
         # The materialized objects corresponding to rows that have been
         # consumed.
-        self._objects: List[AnyModel] = []
+        self._objects: list[AnyModel] = []
 
     def _get_objects(self) -> Iterator[AnyModel]:
         """Construct and generate Model objects for they query. The
@@ -852,7 +836,7 @@ class Results(Generic[AnyModel]):
 
     def _get_indexed_flex_attrs(self) -> Mapping:
         """Index flexible attributes by the entity id they belong to"""
-        flex_values: Dict[int, Dict[str, Any]] = {}
+        flex_values: dict[int, dict[str, Any]] = {}
         for row in self.flex_rows:
             if row["entity_id"] not in flex_values:
                 flex_values[row["entity_id"]] = {}
@@ -861,7 +845,7 @@ class Results(Generic[AnyModel]):
 
         return flex_values
 
-    def _make_model(self, row, flex_values: Dict = {}) -> AnyModel:
+    def _make_model(self, row, flex_values: dict = {}) -> AnyModel:
         """Create a Model object for the given row"""
         cols = dict(row)
         values = {k: v for (k, v) in cols.items() if not k[:4] == "flex"}
@@ -912,7 +896,7 @@ class Results(Generic[AnyModel]):
         except StopIteration:
             raise IndexError(f"result index {n} out of range")
 
-    def get(self) -> Optional[AnyModel]:
+    def get(self) -> AnyModel | None:
         """Return the first matching object, or None if no objects
         match.
         """
@@ -933,10 +917,10 @@ class Transaction:
     current transaction.
     """
 
-    def __init__(self, db: "Database"):
+    def __init__(self, db: Database):
         self.db = db
 
-    def __enter__(self) -> "Transaction":
+    def __enter__(self) -> Transaction:
         """Begin a transaction. This transaction may be created while
         another is active in a different thread.
         """
@@ -951,7 +935,7 @@ class Transaction:
 
     def __exit__(
         self,
-        exc_type: Type[Exception],
+        exc_type: type[Exception],
         exc_value: Exception,
         traceback: TracebackType,
     ):
@@ -970,7 +954,7 @@ class Transaction:
             self._mutated = False
             self.db._db_lock.release()
 
-    def query(self, statement: str, subvals: Sequence = ()) -> List:
+    def query(self, statement: str, subvals: Sequence = ()) -> list:
         """Execute an SQL statement with substitution values and return
         a list of rows from the database.
         """
@@ -1010,7 +994,7 @@ class Database:
     the backend.
     """
 
-    _models: Sequence[Type[Model]] = ()
+    _models: Sequence[type[Model]] = ()
     """The Model subclasses representing tables in this database.
     """
 
@@ -1031,9 +1015,9 @@ class Database:
         self.path = path
         self.timeout = timeout
 
-        self._connections: Dict[int, sqlite3.Connection] = {}
-        self._tx_stacks: DefaultDict[int, List[Transaction]] = defaultdict(list)
-        self._extensions: List[str] = []
+        self._connections: dict[int, sqlite3.Connection] = {}
+        self._tx_stacks: defaultdict[int, list[Transaction]] = defaultdict(list)
+        self._extensions: list[str] = []
 
         # A lock to protect the _connections and _tx_stacks maps, which
         # both map thread IDs to private resources.
@@ -1110,7 +1094,7 @@ class Database:
                 value = value.decode()
             return re.search(pattern, str(value)) is not None
 
-        def bytelower(bytestring: Optional[AnyStr]) -> Optional[AnyStr]:
+        def bytelower(bytestring: AnyStr | None) -> AnyStr | None:
             """A custom ``bytelower`` sqlite function so we can compare
             bytestrings in a semi case insensitive fashion.
 
@@ -1138,7 +1122,7 @@ class Database:
                 conn.close()
 
     @contextlib.contextmanager
-    def _tx_stack(self) -> Generator[List, None, None]:
+    def _tx_stack(self) -> Generator[list]:
         """A context manager providing access to the current thread's
         transaction stack. The context manager synchronizes access to
         the stack map. Transactions should never migrate across threads.
@@ -1231,9 +1215,9 @@ class Database:
 
     def _fetch(
         self,
-        model_cls: Type[AnyModel],
-        query: Optional[Query] = None,
-        sort: Optional[Sort] = None,
+        model_cls: type[AnyModel],
+        query: Query | None = None,
+        sort: Sort | None = None,
     ) -> Results[AnyModel]:
         """Fetch the objects of type `model_cls` matching the given
         query. The query may be given as a string, string sequence, a
@@ -1289,9 +1273,9 @@ class Database:
 
     def _get(
         self,
-        model_cls: Type[AnyModel],
+        model_cls: type[AnyModel],
         id,
-    ) -> Optional[AnyModel]:
+    ) -> AnyModel | None:
         """Get a Model object by its id or None if the id does not
         exist.
         """
