@@ -14,9 +14,11 @@
 
 """Representation of type information for DBCore model fields."""
 
+from __future__ import annotations
+
 import typing
 from abc import ABC
-from typing import Any, Generic, List, TypeVar, Union, cast
+from typing import Any, Generic, TypeVar, cast
 
 from beets.util import str2bool
 
@@ -49,11 +51,11 @@ class Type(ABC, Generic[T, N]):
     """The SQLite column type for the value.
     """
 
-    query: typing.Type[FieldQuery] = SubstringQuery
+    query: type[FieldQuery] = SubstringQuery
     """The `Query` subclass to be used when querying the field.
     """
 
-    model_type: typing.Type[T]
+    model_type: type[T]
     """The Python type that is used to represent the value in the model.
 
     The model is guaranteed to return a value of this type if the field
@@ -69,7 +71,7 @@ class Type(ABC, Generic[T, N]):
         # have a field null_type similar to `model_type` and use that here.
         return cast(N, self.model_type())
 
-    def format(self, value: Union[N, T]) -> str:
+    def format(self, value: N | T) -> str:
         """Given a value of this type, produce a Unicode string
         representing the value. This is used in template evaluation.
         """
@@ -83,7 +85,7 @@ class Type(ABC, Generic[T, N]):
         else:
             return str(value)
 
-    def parse(self, string: str) -> Union[T, N]:
+    def parse(self, string: str) -> T | N:
         """Parse a (possibly human-written) string and return the
         indicated value of this type.
         """
@@ -92,7 +94,7 @@ class Type(ABC, Generic[T, N]):
         except ValueError:
             return self.null
 
-    def normalize(self, value: Any) -> Union[T, N]:
+    def normalize(self, value: Any) -> T | N:
         """Given a value that will be assigned into a field of this
         type, normalize the value to have the appropriate type. This
         base implementation only reinterprets `None`.
@@ -107,8 +109,8 @@ class Type(ABC, Generic[T, N]):
 
     def from_sql(
         self,
-        sql_value: Union[None, int, float, str, bytes],
-    ) -> Union[T, N]:
+        sql_value: None | int | float | str | bytes,
+    ) -> T | N:
         """Receives the value stored in the SQL backend and return the
         value to be stored in the model.
 
@@ -129,7 +131,7 @@ class Type(ABC, Generic[T, N]):
         else:
             return self.normalize(sql_value)
 
-    def to_sql(self, model_value: Any) -> Union[None, int, float, str, bytes]:
+    def to_sql(self, model_value: Any) -> None | int | float | str | bytes:
         """Convert a value as stored in the model object to a value used
         by the database adapter.
         """
@@ -154,7 +156,7 @@ class BaseInteger(Type[int, N]):
     query = NumericQuery
     model_type = int
 
-    def normalize(self, value: Any) -> Union[int, N]:
+    def normalize(self, value: Any) -> int | N:
         try:
             return self.model_type(round(float(value)))
         except ValueError:
@@ -183,7 +185,7 @@ class BasePaddedInt(BaseInteger[N]):
     def __init__(self, digits: int):
         self.digits = digits
 
-    def format(self, value: Union[int, N]) -> str:
+    def format(self, value: int | N) -> str:
         return "{0:0{1}d}".format(value or 0, self.digits)
 
 
@@ -232,13 +234,13 @@ class BaseFloat(Type[float, N]):
     """
 
     sql = "REAL"
-    query: typing.Type[FieldQuery[Any]] = NumericQuery
+    query: type[FieldQuery[Any]] = NumericQuery
     model_type = float
 
     def __init__(self, digits: int = 1):
         self.digits = digits
 
-    def format(self, value: Union[float, N]) -> str:
+    def format(self, value: float | N) -> str:
         return "{0:.{1}f}".format(value or 0, self.digits)
 
 
@@ -264,7 +266,7 @@ class BaseString(Type[T, N]):
     sql = "TEXT"
     query = SubstringQuery
 
-    def normalize(self, value: Any) -> Union[T, N]:
+    def normalize(self, value: Any) -> T | N:
         if value is None:
             return self.null
         else:
@@ -277,7 +279,7 @@ class String(BaseString[str, Any]):
     model_type = str
 
 
-class DelimitedString(BaseString[List[str], List[str]]):
+class DelimitedString(BaseString[list[str], list[str]]):
     """A list of Unicode strings, represented in-database by a single string
     containing delimiter-separated values.
     """
@@ -287,7 +289,7 @@ class DelimitedString(BaseString[List[str], List[str]]):
     def __init__(self, delimiter: str):
         self.delimiter = delimiter
 
-    def format(self, value: List[str]):
+    def format(self, value: list[str]):
         return self.delimiter.join(value)
 
     def parse(self, string: str):
@@ -295,7 +297,7 @@ class DelimitedString(BaseString[List[str], List[str]]):
             return []
         return string.split(self.delimiter)
 
-    def to_sql(self, model_value: List[str]):
+    def to_sql(self, model_value: list[str]):
         return self.delimiter.join(model_value)
 
 
