@@ -15,14 +15,17 @@
 
 from __future__ import annotations
 
-from typing import Iterable
+from collections.abc import Iterable
+from typing import TYPE_CHECKING
 
 import librosa
 
-from beets.importer import ImportTask
-from beets.library import Item, Library
 from beets.plugins import BeetsPlugin
 from beets.ui import Subcommand, should_write
+
+if TYPE_CHECKING:
+    from beets.importer import ImportTask
+    from beets.library import Item, Library
 
 
 class AutoBPMPlugin(BeetsPlugin):
